requirements.txt
boto==2.42.0
dj-database-url==0.4.1
Django==1.9.8
django-cors-headers==1.1.0
django-debug-toolbar==1.5
django-forms-bootstrap==3.0.1
django-storages==1.5.0
djangorestframework==3.4.0
gunicorn==19.4.0
Pillow==2.9.0
sqlparse==0.2.0
MySQL-python==1.2.5